+++ scripts/basic/buy_only_three_times_example.py
