+++ scripts/basic/format_status_example.py
