+++ scripts/basic/log_price_example.py
