+++ scripts/community/buy_dip_example.py
