+++ scripts/community/directional_strategy_macd_bb.py
