+++ scripts/community/directional_strategy_rsi_spot.py
