+++ scripts/community/directional_strategy_trend_follower.py
