+++ scripts/community/directional_strategy_widening_ema_bands.py
