+++ scripts/community/fixed_grid.py
