+++ scripts/community/macd_bb_directional_strategy.py
