+++ scripts/community/simple_arbitrage_example.py
