+++ scripts/community/simple_rsi_no_config.py
