+++ scripts/community/simple_vwap_no_config.py
