+++ scripts/v2_funding_rate_arb.py
