+++ scripts/v2_generic_with_cash_out.py
